requirements.txt
folium~=0.12.1
requests~=2.25.1
Flask~=1.1.2
geopy~=2.1.0
python-dotenv==0.15.0